Any hit using BVH
